requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Pillow
numpy 
transformers
torch

